+++ packages/ui/CHANGELOG.md
@@ -1,4 +1,16 @@
-# @process-optimizer-frontend/ui
+# @boostv/process-optimizer-frontend-ui
+
+## 1.0.2
+
+### Patch Changes
+
+- 4966eef: Add remaining theme libraries
+
+## 1.0.1
+
+### Patch Changes
+
+- ae4b2bc: Convert some peer dependencies into actual dependencies
 
 ## 1.0.0
 
@@ -6,5 +18,5 @@
 
 - First monorepo release
 - Updated dependencies
-  - @process-optimizer-frontend/plots@1.0.0
-  - @process-optimizer-frontend/core@1.0.0
+  - @boostv/process-optimizer-frontend-plots@1.0.0
+  - @boostv/process-optimizer-frontend-core@1.0.0